fraudfinder/vertex_ai/05-v02_custom/source/requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
xgboost
pandas
sklearn
google-cloud-storage
